Select language mode based on reference class
