+++ core/src/main/kotlin/edu/illinois/cs/cs125/answerable/LanguageMode.kt
@@ -1,5 +1,7 @@
 package edu.illinois.cs.cs125.answerable
 
+import java.util.*
+import kotlin.math.min
 import java.lang.reflect.Array as ReflectArray
 
 internal interface LanguageMode {
@@ -64,6 +66,106 @@ private val defaultPrimitiveEdgeCases = mapOf(
         Boolean::class.java to ArrayWrapper(booleanArrayOf())
 )
 
+internal val defaultIntGen = object : Gen<Int> {
+    override fun generate(complexity: Int, random: Random): Int {
+        var comp = complexity
+        if (complexity > Int.MAX_VALUE / 2) {
+            comp = Int.MAX_VALUE / 2
+        }
+        return random.nextInt(comp * 2 + 1) - comp
+    }
+}
+
+internal val defaultDoubleGen = object : Gen<Double> {
+    override fun generate(complexity: Int, random: Random): Double {
+        val denom = random.nextDouble() * (1e10 - 1) + 1
+        val num = (random.nextDouble() * 2 * complexity * denom) - complexity * denom
+        return num / denom
+    }
+}
+
+internal val defaultFloatGen = object : Gen<Float> {
+    override fun generate(complexity: Int, random: Random): Float {
+        val denom = random.nextDouble() * (1e10 - 1) + 1
+        val num = (random.nextDouble() * 2 * complexity * denom) - complexity * denom
+        return (num / denom).toFloat() // if complexity is > 1e38, this stops being uniform
+    }
+}
+
+internal val defaultByteGen = object : Gen<Byte> {
+    override fun generate(complexity: Int, random: Random): Byte {
+        return (random.nextInt(complexity * 2 + 1) - complexity).toByte()
+    }
+}
+
+internal val defaultShortGen = object : Gen<Short> {
+    override fun generate(complexity: Int, random: Random): Short {
+        return (random.nextInt(complexity * 2 + 1) - complexity).toShort()
+    }
+}
+
+internal val defaultLongGen = object : Gen<Long> {
+    // see Random.nextInt(int) algorithm.
+    private fun Random.nextLong(bound: Long): Long {
+        var bits: Long
+        var value: Long
+        do {
+            bits = (nextLong() shl 1) shr 1
+            value = bits % bound
+        } while (bits - value + (bound - 1) < 0L)
+        return value
+    }
+
+    override fun generate(complexity: Int, random: Random): Long {
+        return random.nextLong(complexity.toLong() * 4 + 1) - (complexity.toLong() * 2)
+    }
+}
+
+internal val defaultCharGen = object : Gen<Char> {
+    private fun Char.isPrintableAscii(): Boolean = this.toInt() in 32..126
+
+    private fun Char.isPrint(): Boolean = isPrintableAscii() || Character.UnicodeBlock.of(this) in setOf(
+            Character.UnicodeBlock.CYRILLIC, Character.UnicodeBlock.CYRILLIC_SUPPLEMENTARY, Character.UnicodeBlock.TAMIL, 
+            Character.UnicodeBlock.CURRENCY_SYMBOLS, Character.UnicodeBlock.ARROWS, Character.UnicodeBlock.SUPPLEMENTAL_ARROWS_A,
+            Character.UnicodeBlock.ETHIOPIC_EXTENDED, Character.UnicodeBlock.CJK_RADICALS_SUPPLEMENT, 
+            Character.UnicodeBlock.KANGXI_RADICALS, Character.UnicodeBlock.KATAKANA_PHONETIC_EXTENSIONS,
+            Character.UnicodeBlock.ENCLOSED_CJK_LETTERS_AND_MONTHS, Character.UnicodeBlock.OLD_PERSIAN
+    )
+
+    override fun generate(complexity: Int, random: Random): Char {
+        return if (random.nextDouble() < min(.15/32 * complexity, .15)) {
+            var char: Char
+            do {
+                char = random.nextInt(0x10000).toChar()
+            } while (!char.isPrint())
+            char
+        } else {
+            (random.nextInt(95) + 32).toChar()
+        }
+    }
+}
+
+internal val defaultAsciiGen = object : Gen<Char> {
+    override fun generate(complexity: Int, random: Random): Char {
+        return (random.nextInt(95) + 32).toChar()
+    }
+}
+
+internal val defaultBooleanGen = object : Gen<Boolean> {
+    override fun generate(complexity: Int, random: Random): Boolean = random.nextBoolean()
+}
+
+private val primitiveGenerators = mapOf(
+        Int::class.java     to defaultIntGen,
+        Double::class.java  to defaultDoubleGen,
+        Float::class.java   to defaultFloatGen,
+        Byte::class.java    to defaultByteGen,
+        Short::class.java   to defaultShortGen,
+        Long::class.java    to defaultLongGen,
+        Char::class.java    to defaultCharGen,
+        Boolean::class.java to defaultBooleanGen
+)
+
 internal object JavaMode : LanguageMode {
     override val defaultSimpleCases: Map<Class<*>, ArrayWrapper>
         get() = valueSimpleCases + arraySimpleCases
@@ -75,7 +177,7 @@ internal object JavaMode : LanguageMode {
             }
         }
     override val defaultGenerators: Map<Class<*>, Gen<*>>
-        get() = TODO("not implemented") //To change initializer of created properties use File | Settings | File Templates.
+        get() = primitiveGenerators
 }
 
 internal object KotlinMode : LanguageMode {
@@ -89,5 +191,9 @@ internal object KotlinMode : LanguageMode {
             }
         }
     override val defaultGenerators: Map<Class<*>, Gen<*>>
-        get() = TODO("not implemented") //To change initializer of created properties use File | Settings | File Templates.
+        get() = primitiveGenerators
+}
+
+internal fun getLanguageMode(clazz: Class<*>): LanguageMode {
+    return if (clazz.isAnnotationPresent(Metadata::class.java)) KotlinMode else JavaMode
 }
