+++ core/src/main/kotlin/edu/illinois/cs/cs125/answerable/TestGeneration.kt
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.Assertions.*
 import org.opentest4j.AssertionFailedError
 import java.util.*
 import kotlin.math.min
-import java.lang.Character.UnicodeBlock.*
 import java.lang.IllegalStateException
 import java.lang.reflect.*
 import java.lang.reflect.Array as ReflectArray
@@ -44,6 +43,7 @@ class TestGenerator(
     // "Usable" members are from the opened (un-final-ified) mirror of the original reference class.
     // The original members are used for certain checks so a nice class name can be displayed.
 
+    internal val languageMode = getLanguageMode(referenceClass)
     internal val typePool = TypePool(bytecodeProvider,
             if (referenceClass.classLoader == javaClass.classLoader) javaClass.classLoader else referenceClass.classLoader?.parent ?: javaClass.classLoader)
     private val controlClass: Class<*> = getDefiningKotlinFileClass(referenceClass, typePool) ?: referenceClass
@@ -114,7 +114,7 @@ class TestGenerator(
     }
 
     internal fun buildGeneratorMap(random: Random, submittedClassGenerator: Method? = null): Map<Pair<Type, String?>, GenWrapper<*>> {
-        val generatorMapBuilder = GeneratorMapBuilder(params.toSet(), random, typePool, if (isStatic) null else usableReferenceClass)
+        val generatorMapBuilder = GeneratorMapBuilder(params.toSet(), random, typePool, if (isStatic) null else usableReferenceClass, languageMode)
 
         val enabledGens: List<Pair<Pair<Type, String?>, CustomGen>> = usableControlClass.getEnabledGenerators(enabledNames).map {
             return@map if (it.returnType == usableReferenceClass && submittedClassGenerator != null) {
@@ -149,11 +149,11 @@ class TestGenerator(
     }
 
     private fun getEdgeCases(clazz: Class<*>, types: Array<Type>): Map<Type, ArrayWrapper?> {
-        val all = defaultEdgeCases + clazz.getEnabledEdgeCases(enabledNames)
+        val all = languageMode.defaultEdgeCases + clazz.getEnabledEdgeCases(enabledNames)
         return mapOf(*types.map { it to all[it] }.toTypedArray())
     }
     private fun getSimpleCases(clazz: Class<*>, types: Array<Type>): Map<Type, ArrayWrapper?> {
-        val all = defaultSimpleCases + clazz.getEnabledSimpleCases(enabledNames)
+        val all = languageMode.defaultSimpleCases + clazz.getEnabledSimpleCases(enabledNames)
         return mapOf(*types.map { it to all[it] }.toTypedArray())
     }
 
@@ -799,8 +799,10 @@ operator fun <T> MutableMap<Pair<Type, String?>, T>.set(type: Type, newVal: T) {
 // NOTE: [Generator Keys]
 // goalTypes holds types that we need generators for. @UseGenerator annotations allow specifying a specific generator.
 // The string in the Pair is non-null iff a specific generator is requested.
-private class GeneratorMapBuilder(goalTypes: Collection<Pair<Type, String?>>, private val random: Random, private val pool: TypePool, private val receiverType: Class<*>?) {
+private class GeneratorMapBuilder(goalTypes: Collection<Pair<Type, String?>>, private val random: Random, private val pool: TypePool,
+                                  private val receiverType: Class<*>?, private val languageMode: LanguageMode) {
     private var knownGenerators: MutableMap<Pair<Type, String?>, Lazy<Gen<*>>> = mutableMapOf()
+    private val defaultGenerators: Map<Pair<Class<*>, String?>, Gen<*>> = languageMode.defaultGenerators.mapKeys { (k, _) -> Pair(k, null) }
     init {
         defaultGenerators.forEach { (k, v) -> accept(k, v) }
         knownGenerators[String::class.java] = lazy { DefaultStringGen(knownGenerators[Char::class.java]!!.value) }
@@ -895,18 +897,6 @@ private class GeneratorMapBuilder(goalTypes: Collection<Pair<Type, String?>>, pr
         return discovered
     }
 
-    companion object {
-        private val defaultGenerators: Map<Pair<Class<*>, String?>, Gen<*>> = mapOf(
-            Int::class.java     to defaultIntGen,
-            Double::class.java  to defaultDoubleGen,
-            Float::class.java   to defaultFloatGen,
-            Byte::class.java    to defaultByteGen,
-            Short::class.java   to defaultShortGen,
-            Long::class.java    to defaultLongGen,
-            Char::class.java    to defaultCharGen,
-            Boolean::class.java to defaultBooleanGen
-        ).mapKeys { (k, _) -> Pair(k, null) }
-    }
 }
 
 internal class GenWrapper<T>(val gen: Gen<T>, private val random: Random) {
@@ -926,89 +916,6 @@ internal class CustomGen(private val gen: Method) : Gen<Any?> {
     override fun generate(complexity: Int, random: Random): Any? = gen(null, complexity, random)
 }
 
-internal val defaultIntGen = object : Gen<Int> {
-    override fun generate(complexity: Int, random: Random): Int {
-        var comp = complexity
-        if (complexity > Int.MAX_VALUE / 2) {
-            comp = Int.MAX_VALUE / 2
-        }
-        return random.nextInt(comp * 2 + 1) - comp
-    }
-}
-
-internal val defaultDoubleGen = object : Gen<Double> {
-    override fun generate(complexity: Int, random: Random): Double {
-        val denom = random.nextDouble() * (1e10 - 1) + 1
-        val num = (random.nextDouble() * 2 * complexity * denom) - complexity * denom
-        return num / denom
-    }
-}
-
-internal val defaultFloatGen = object : Gen<Float> {
-    override fun generate(complexity: Int, random: Random): Float {
-        val denom = random.nextDouble() * (1e10 - 1) + 1
-        val num = (random.nextDouble() * 2 * complexity * denom) - complexity * denom
-        return (num / denom).toFloat() // if complexity is > 1e38, this stops being uniform
-    }
-}
-
-internal val defaultByteGen = object : Gen<Byte> {
-    override fun generate(complexity: Int, random: Random): Byte {
-        return (random.nextInt(complexity * 2 + 1) - complexity).toByte()
-    }
-}
-
-internal val defaultShortGen = object : Gen<Short> {
-    override fun generate(complexity: Int, random: Random): Short {
-        return (random.nextInt(complexity * 2 + 1) - complexity).toShort()
-    }
-}
-
-internal val defaultLongGen = object : Gen<Long> {
-    // see Random.nextInt(int) algorithm.
-    private fun Random.nextLong(bound: Long): Long {
-        var bits: Long
-        var value: Long
-        do {
-            bits = (nextLong() shl 1) shr 1
-            value = bits % bound
-        } while (bits - value + (bound - 1) < 0L)
-        return value
-    }
-
-    override fun generate(complexity: Int, random: Random): Long {
-        return random.nextLong(complexity.toLong() * 4 + 1) - (complexity.toLong() * 2)
-    }
-}
-
-internal val defaultCharGen = object : Gen<Char> {
-    private fun Char.isPrintableAscii(): Boolean = this.toInt() in 32..126
-
-    private fun Char.isPrint(): Boolean = isPrintableAscii() || of(this) in setOf(
-        CYRILLIC, CYRILLIC_SUPPLEMENTARY, TAMIL, CURRENCY_SYMBOLS, ARROWS, SUPPLEMENTAL_ARROWS_A,
-        ETHIOPIC_EXTENDED, CJK_RADICALS_SUPPLEMENT, KANGXI_RADICALS, KATAKANA_PHONETIC_EXTENSIONS,
-        ENCLOSED_CJK_LETTERS_AND_MONTHS, OLD_PERSIAN
-    )
-
-    override fun generate(complexity: Int, random: Random): Char {
-        return if (random.nextDouble() < min(.15/32 * complexity, .15)) {
-            var char: Char
-            do {
-                char = random.nextInt(0x10000).toChar()
-            } while (!char.isPrint())
-            char
-        } else {
-            (random.nextInt(95) + 32).toChar()
-        }
-    }
-}
-
-internal val defaultAsciiGen = object : Gen<Char> {
-    override fun generate(complexity: Int, random: Random): Char {
-        return (random.nextInt(95) + 32).toChar()
-    }
-}
-
 internal class DefaultStringGen(private val cGen: Gen<*>) : Gen<String> {
     override fun generate(complexity: Int, random: Random): String {
         val len = random.nextInt(complexity + 1)
@@ -1017,16 +924,24 @@ internal class DefaultStringGen(private val cGen: Gen<*>) : Gen<String> {
     }
 }
 
-internal val defaultBooleanGen = object : Gen<Boolean> {
-    override fun generate(complexity: Int, random: Random): Boolean = random.nextBoolean()
-}
-
 internal class DefaultArrayGen<T>(private val tGen: Gen<T>, private val tClass: Class<*>) : Gen<Any> {
     override fun generate(complexity: Int, random: Random): Any {
         return ReflectArray.newInstance(tClass, random.nextInt(complexity + 1)).also {
             val wrapper = ArrayWrapper(it)
             (0 until wrapper.size).forEach { idx -> wrapper[idx] = tGen(random.nextInt(complexity + 1), random) }
         }
+    }
+}
+
+internal class DefaultListGen<T>(private val tGen: Gen<T>) : Gen<List<T>> {
+    override fun generate(complexity: Int, random: Random): List<T> {
+        fun genList(complexity: Int, length: Int): List<T> =
+                if (length <= 0) {
+                    listOf()
+                } else {
+                    listOf(tGen(random.nextInt(complexity + 1), random)) + genList(complexity, length - 1)
+                }
+        return genList(complexity, random.nextInt(complexity + 1))
     }
 }
 
@@ -1043,18 +958,6 @@ internal class ArrayWrapper(val array: Any) {
     }
 }
 internal fun ArrayWrapper?.isNullOrEmpty() = this == null || this.size == 0
-
-internal class DefaultListGen<T>(private val tGen: Gen<T>) : Gen<List<T>> {
-    override fun generate(complexity: Int, random: Random): List<T> {
-        fun genList(complexity: Int, length: Int): List<T> =
-            if (length <= 0) {
-                listOf()
-            } else {
-                listOf(tGen(random.nextInt(complexity + 1), random)) + genList(complexity, length - 1)
-            }
-        return genList(complexity, random.nextInt(complexity + 1))
-    }
-}
 
 /**
  * The types of behaviors that methods under test can have.
